working ish, term and some more commands work
