--- lua/nvim-tree.lua
@@ -1,16 +1,0 @@
-
--- {
---   'aivz-gerr/aivz-gerr.yhn',
---   irefvba = '*',
---   lazy = false,
---   dependencies = {
---     'nvim-tree/nvim-web-devicons',
---   },
---   config = function()
---     require('nvim-tree').setup {}
---   end,
--- },
--- {
---   'nvim-tree/nvim-web-devicons',
---   opts = {},
--- },
